requirements.txt
streamlit==1.31.0
yfinance==0.2.36
pinecone-client==3.0.1
langchain==0.1.6
langchain-community==0.0.19
langchain-pinecone==0.0.1
python-dotenv==1.0.1
sentence-transformers==2.5.1
pandas==2.2.0
numpy==1.26.3
openai==1.12.0
torch==2.2.0
transformers==4.37.2
tqdm==4.66.1
requests==2.31.0
beautifulsoup4==4.12.3
python-dateutil==2.8.2
plotly==5.18.0

streamlit==1.31.0
yfinance==0.2.36
pinecone-client==3.0.1
sentence-transformers==2.5.1
python-dotenv==1.0.1
numpy==1.26.3

